+++ src/gmprocess/metrics/combination/quadratic_mean.py
@@ -38,7 +38,7 @@ class Quadratic_Mean(Combination):
                     "spectra": np.sqrt((h1["spectra"] ** 2 + h2["spectra"] ** 2) / 2),
                 }
             else:
-                qm = {"": np.sqrt(np.mean([h1 ** 2, h2 ** 2]))}
+                qm = {"": np.sqrt(np.mean([h1**2, h2**2]))}
         else:
             # Just for tests?
             horizontals = self.combination_data