+++ tests/gmprocess/waveform_processing/nn_quality_assurance_test.py
@@ -19,7 +19,6 @@ setup_logger()
 
 
 def test_nnet():
-
     conf = get_config()
 
     update = {