+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# main
+ - Bugfixes
+   - Config wasn't being passed to COSMOS reader, or to StreamArray.
+
 ## 1.2.5 / 2023-03-17
  - Updated obspy data reader to allow station XML base names to match corresponding miniseed file names.
  - Added support for passing in a config directory to `gmrecords config -u`